Make changes to build process to respect dependancies.
When we use this for real, we would like to have the following workflow:

* run `fig run protodeps`, which calls tasks in protodeps.rake
* Copy all protobufs everywhere in the site into a single directory
* Run protoc to generate .pb.cc and .fastproto.cpp files from .proto
files
* Copy the static bits (compiler/runtime) into this test dir
* Run `rake` to build the .so file output

If we do things like this, it means we can make n-many invocations of
protoc, generating a single {.pb.cc,.pb.h,.fastproto.cpp,.fastproto.h}
set out of a single protobuf file at a time, instead of passing all
protobuf files onto a single command line for protoc.

The benefit of this is that, when a proto file is updated, we can use
protoc to only update the corresponding generated files without dirtying
the other ones. Then, in theory, make should be able to re-compile only
those protobuf classes, and re-lik them with already generated object
files.

Unfortunately, it's not so simple. If a header file is changed, in
theory every file that transitively depends on that header file needs
to be recompiled. For this reason, mkmf generates a rule that recompiles
everything when a header file is updated. The protobuf compiler always
updates the header, so....